requirements.txt
altgraph==0.17.3
anyio==3.6.2
appnope==0.1.3
APScheduler==3.10.1
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
googlesearch-python==1.2.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jupyter_client==8.1.0
jupyter_core==5.3.0
lxml==4.9.2
macholib==1.16.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.3
openpyxl==3.1.2
packaging==23.0
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.4
python-dateutil==2.8.2
python-telegram-bot==20.2
pytz==2023.3
pyzmq==25.0.2
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
tk==0.1.0
tornado==6.2
traitlets==5.9.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
wcwidth==0.2.6
